+++ packages/common/src/api/index.ts
@@ -24,7 +24,6 @@ export * from './tan-query/useFollowing'
 export * from './tan-query/useHandleInUse'
 export * from './tan-query/useHandleReservedStatus'
 export * from './tan-query/useMutualFollowers'
-export * from './tan-query/useNotifications'
 export * from './tan-query/usePurchases'
 export * from './tan-query/usePurchasesCount'
 export * from './tan-query/useRelatedArtists'
