+++ packages/common/src/api/tan-query/useMutualFollowers.ts
@@ -1,15 +1,10 @@
-import {
-  InfiniteData,
-  useInfiniteQuery,
-  useQueryClient
-} from '@tanstack/react-query'
+import { useInfiniteQuery, useQueryClient } from '@tanstack/react-query'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { userMetadataListFromSDK } from '~/adapters/user'
 import { useAudiusQueryContext } from '~/audius-query'
-import { ID, OptionalId } from '~/models'
+import { ID, OptionalId, UserMetadata } from '~/models'
 import { Id } from '~/models/Identifiers'
-import { User } from '~/models/User'
 import { getUserId } from '~/store/account/selectors'
 
 import { QUERY_KEYS } from './queryKeys'
@@ -35,7 +30,7 @@ export const useMutualFollowers = (
   return useInfiniteQuery({
     queryKey: [QUERY_KEYS.mutualFollowers, userId, pageSize],
     initialPageParam: 0,
-    getNextPageParam: (lastPage, allPages) => {
+    getNextPageParam: (lastPage: UserMetadata[], allPages) => {
       if (lastPage.length < pageSize) return undefined
       return allPages.length * pageSize
     },
@@ -51,7 +46,7 @@ export const useMutualFollowers = (
       primeUserData({ users, queryClient, dispatch })
       return users
     },
-    select: (data: InfiniteData<User[]>) => data.pages.flat(),
+    select: (data) => data.pages.flat(),
     staleTime: config?.staleTime,
     enabled: config?.enabled !== false && !!userId && !!currentUserId
   })
