+++ packages/mobile/src/screens/user-list-screen/FollowersScreen.tsx
@@ -6,8 +6,8 @@ import { useProfileRoute } from 'app/hooks/useRoute'
 
 import { ChatBlastWithAudienceCTA } from '../chat-screen/ChatBlastWithAudienceCTA'
 
+import { UserList } from './UserList'
 import { UserListScreen } from './UserListScreen'
-import { UserListV2 } from './UserListV2'
 
 const messages = {
   title: 'Followers'
@@ -23,7 +23,7 @@ export const FollowersScreen = () => {
   return (
     <UserListScreen title={messages.title} titleIcon={IconUserFollowers}>
       <>
-        <UserListV2 {...query} tag='FOLLOWERS' />
+        <UserList {...query} tag='FOLLOWERS' />
         {currentUserId === userId ? (
           <Box w='100%' style={{ position: 'absolute', bottom: 0 }}>
             <ChatBlastWithAudienceCTA audience={ChatBlastAudience.FOLLOWERS} />
