+++ packages/mobile/src/screens/user-list-screen/SupportingUsersScreen.tsx
@@ -3,8 +3,8 @@ import { useSupportedUsers } from '@audius/common/api'
 import { IconTipping } from '@audius/harmony-native'
 import { useRoute } from 'app/hooks/useRoute'
 
+import { UserList } from './UserList'
 import { UserListScreen } from './UserListScreen'
-import { UserListV2 } from './UserListV2'
 
 const messages = {
   title: 'Supporting'
@@ -17,7 +17,7 @@ export const SupportingUsersScreen = () => {
 
   return (
     <UserListScreen title={messages.title} titleIcon={IconTipping}>
-      <UserListV2
+      <UserList
         {...query}
         data={query.data?.map((supporter) => supporter.receiver)}
         tag='SUPPORTING'
