+++ packages/web/src/app/web-player/WebPlayer.jsx
@@ -67,7 +67,7 @@ import FollowersPage from 'pages/followers-page/FollowersPage'
 import FollowingPage from 'pages/following-page/FollowingPage'
 import HistoryPage from 'pages/history-page/HistoryPage'
 import { NotFoundPage } from 'pages/not-found-page/NotFoundPage'
-import NotificationUsersPage from 'pages/notification-users-page/NotificationUsersPage'
+import { NotificationUsersPage } from 'pages/notification-users-page/NotificationUsersPage'
 import { PayAndEarnPage } from 'pages/pay-and-earn-page/PayAndEarnPage'
 import { TableType } from 'pages/pay-and-earn-page/types'
 import { PremiumTracksPage } from 'pages/premium-tracks-page/PremiumTracksPage'
