--- packages/web/src/components/user-list-modal/components/UserListModal.module.css
@@ -1,53 +1,0 @@
-.modalBody {
-  max-width: 560px;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-.modalHeader {
-  margin: 0px 16px;
-  border-bottom-color: var(--harmony-n-150);
-}
-
-.modalTitle {
-  font-weight: var(--harmony-font-heavy);
-  letter-spacing: 0.5px;
-  color: var(--harmony-n-400);
-  font-size: var(--harmony-font-l);
-  text-transform: uppercase;
-}
-
-.titleContainer {
-  display: flex;
-  align-items: center;
-}
-
-.icon {
-  margin-right: 10px;
-  width: 24px;
-  height: 24px;
-}
-
-.icon path {
-  fill: var(--harmony-n-400);
-}
-
-.scrollable {
-  min-height: 0;
-  max-height: 690px;
-  height: 100%;
-  overflow-y: auto;
-  overflow-x: hidden;
-}
-
-.titleNameContainer {
-  display: flex;
-  align-items: center;
-}
-
-.titleName {
-  max-width: 240px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
