--- packages/web/src/components/user-list/UserListV2.module.css
@@ -1,40 +1,0 @@
-.content {
-  height: 100%;
-  position: relative;
-}
-
-.userContainer {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: var(--unit-2);
-}
-
-.notLastUser {
-  border-bottom: 1px solid var(--neutral-light-8);
-}
-
-.spacer {
-  height: var(--unit-32);
-}
-
-.loadingAnimation {
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  opacity: 0;
-  transition: opacity 0.1s ease-in-out;
-  display: flex;
-  justify-content: center;
-  padding: var(--unit-4);
-}
-
-.loadingAnimation.show {
-  opacity: 1;
-}
-
-.spinner {
-  height: var(--unit-8);
-  width: var(--unit-8);
-}
