+++ packages/web/src/components/user-list/lists/FollowersUserList.tsx
@@ -2,11 +2,11 @@ import { useFollowers } from '@audius/common/api'
 import { followersUserListSelectors } from '@audius/common/store'
 import { useSelector } from 'react-redux'
 
-import { UserListV2 } from 'components/user-list/UserListV2'
+import { UserList } from 'components/user-list/UserList'
 
 export const FollowersUserList = () => {
   const userId = useSelector(followersUserListSelectors.getId)
   const query = useFollowers({ userId })
 
-  return <UserListV2 {...query} />
+  return <UserList {...query} />
 }
