+++ packages/web/src/components/user-list/lists/FollowingUserList.tsx
@@ -2,11 +2,11 @@ import { useFollowing } from '@audius/common/api'
 import { followingUserListSelectors } from '@audius/common/store'
 import { useSelector } from 'react-redux'
 
-import { UserListV2 } from 'components/user-list/UserListV2'
+import { UserList } from 'components/user-list/UserList'
 
 export const FollowingUserList = () => {
   const userId = useSelector(followingUserListSelectors.getId)
   const query = useFollowing({ userId })
 
-  return <UserListV2 {...query} />
+  return <UserList {...query} />
 }
