+++ packages/web/src/components/user-list/lists/MutualsUserList.tsx
@@ -2,11 +2,11 @@ import { useMutualFollowers } from '@audius/common/api'
 import { mutualsUserListSelectors } from '@audius/common/store'
 import { useSelector } from 'react-redux'
 
-import { UserListV2 } from 'components/user-list/UserListV2'
+import { UserList } from 'components/user-list/UserList'
 
 export const MutualsUserList = () => {
   const userId = useSelector(mutualsUserListSelectors.getId)
   const query = useMutualFollowers({ userId })
 
-  return <UserListV2 {...query} />
+  return <UserList {...query} />
 }
