+++ packages/web/src/components/user-list/lists/RelatedArtistsUserList.tsx
@@ -2,11 +2,11 @@ import { useRelatedArtists } from '@audius/common/api'
 import { relatedArtistsUserListSelectors } from '@audius/common/store'
 import { useSelector } from 'react-redux'
 
-import { UserListV2 } from 'components/user-list/UserListV2'
+import { UserList } from 'components/user-list/UserList'
 
 export const RelatedArtistsUserList = () => {
   const userId = useSelector(relatedArtistsUserListSelectors.getId)
   const query = useRelatedArtists({ artistId: userId })
 
-  return <UserListV2 {...query} />
+  return <UserList {...query} />
 }
