+++ packages/web/src/components/user-list/lists/SupportingUserList.tsx
@@ -2,14 +2,14 @@ import { useSupportedUsers } from '@audius/common/api'
 import { supportingUserListSelectors } from '@audius/common/store'
 import { useSelector } from 'react-redux'
 
-import { UserListV2 } from 'components/user-list/UserListV2'
+import { UserList } from 'components/user-list/UserList'
 
 export const SupportingUserList = () => {
   const userId = useSelector(supportingUserListSelectors.getId)
   const query = useSupportedUsers({ userId })
 
   return (
-    <UserListV2
+    <UserList
       {...query}
       data={query.data?.map((supporter) => supporter.receiver)}
     />
