+++ packages/web/src/pages/modals/Modals.tsx
@@ -44,7 +44,7 @@ import { USDCPurchaseDetailsModal } from 'components/usdc-purchase-details-modal
 import { USDCTransactionDetailsModal } from 'components/usdc-transaction-details-modal/USDCTransactionDetailsModal'
 import TierExplainerModal from 'components/user-badges/TierExplainerModal'
 // import UserListModal from 'components/user-list-modal/ConnectedUserListModal'
-import UserListModal from 'components/user-list-modal/components/UserListModalV2'
+import UserListModal from 'components/user-list-modal/components/UserListModal'
 import { WaitForDownloadModal } from 'components/wait-for-download-modal/WaitForDownloadModal'
 import { WelcomeModal } from 'components/welcome-modal/WelcomeModal'
 import { WithdrawUSDCModal } from 'components/withdraw-usdc-modal/WithdrawUSDCModal'
