+++ packages/web/src/pages/supporting-page/SupportingPage.tsx
@@ -1,15 +1,8 @@
 import { useContext, useEffect } from 'react'
 
-import {
-  supportingUserListSelectors,
-  SUPPORTING_USER_LIST_TAG
-} from '@audius/common/store'
-
 import MobilePageContainer from 'components/mobile-page-container/MobilePageContainer'
 import NavContext, { LeftPreset } from 'components/nav/mobile/NavContext'
-import { UserList } from 'components/user-list/UserList'
-
-const { getUserList } = supportingUserListSelectors
+import { SupportingUserList } from 'components/user-list/lists/SupportingUserList'
 
 const messages = {
   title: 'Supporting'
@@ -26,7 +19,7 @@ const SupportingPage = () => {
 
   return (
     <MobilePageContainer fullHeight>
-      <UserList stateSelector={getUserList} tag={SUPPORTING_USER_LIST_TAG} />
+      <SupportingUserList />
     </MobilePageContainer>
   )
 }
